#1 Refactor : BoardController 하나의 클래스로 병합함
